restructure leave view
